Move JDBC dialects into -lib-db
