+++ tracdap-libs/tracdap-lib-db/src/main/java/org/finos/tracdap/common/db/dialects/IDialect.java
@@ -1,0 +1,35 @@
+/*
+ * Licensed to the Fintech Open Source Foundation (FINOS) under one or
+ * more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * FINOS licenses this file to you under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with the
+ * License. You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.finos.tracdap.common.db.dialects;
+
+import org.finos.tracdap.common.db.JdbcDialect;
+import org.finos.tracdap.common.db.JdbcErrorCode;
+
+import java.sql.SQLException;
+
+
+public interface IDialect {
+
+    JdbcDialect dialectCode();
+
+    JdbcErrorCode mapErrorCode(SQLException e);
+
+    boolean supportsGeneratedKeys();
+
+    int booleanType();
+}
